requirements.txt
flask
waitress
sqlalchemy
mysql-connector-python
passlib
pandas
numpy
flask-caching
pymysql
flask-compress
python-dotenv
werkzeug
markupsafe
openpyxl
xlsxwriter
pyxlsb
pymysql
cryptography
pymysql